move shared fixtures to conftest.py
