+++ tests/scoring/test_data_links.py
@@ -1,14 +1,7 @@
 import pandas as pd
 from pandas.util.testing import assert_frame_equal
 import pytest
-from pytest import fixture
 from nplinker.metabolomics.singleton_family import SingletonFamily
-from nplinker.scoring.linking import DataLinks
-
-
-@fixture(scope='module')
-def datalinks(gcfs, spectra, mfs, strains) -> DataLinks:
-    return DataLinks(gcfs, spectra, mfs, strains)
 
 
 def test_init(datalinks):
