+++ tests/scoring/test_nplinker_scoring.py
@@ -1,60 +1,7 @@
 import numpy as np
 import pytest
-from pytest import fixture
-from nplinker.nplinker import NPLinker
 from nplinker.scoring import LinkCollection
-from nplinker.scoring import MetcalfScoring
 from nplinker.scoring import ObjectLink
-from nplinker.scoring.linking import DataLinks
-from nplinker.scoring.linking import LinkFinder
-from .. import DATA_DIR
-
-
-@fixture(scope='module')
-def datalinks(gcfs, spectra, mfs, strains) -> DataLinks:
-    """DataLinks object. See `test_data_links.py` for its values."""
-    return DataLinks(gcfs, spectra, mfs, strains)
-
-
-@fixture(scope='module')
-def linkfinder(datalinks) -> LinkFinder:
-    """LinkFinder object. See `test_link_finder.py` for its values."""
-    linkfinder = LinkFinder()
-    linkfinder.calc_score(datalinks, link_type='spec-gcf')
-    linkfinder.calc_score(datalinks, link_type='mf-gcf')
-    return linkfinder
-
-
-@fixture(scope='module')
-def npl(gcfs, spectra, mfs, strains, tmp_path_factory) -> NPLinker:
-    """Constructed NPLinker object.
-
-    This NPLinker object does not do loading `npl.load_data()`, instead we
-    manually set its attributes to the values we want to test.
-
-    The config file `nplinker_demo1.toml` does not affect the tests, just
-    making sure the NPLinker object can be created succesfully.
-    """
-    npl = NPLinker(str(DATA_DIR / 'nplinker_demo1.toml'))
-    npl._gcfs = gcfs
-    npl._spectra = spectra
-    npl._molfams = mfs
-    npl._strains = strains
-    npl._gcf_lookup = {gcf.gcf_id: gcf for gcf in gcfs}
-    npl._mf_lookup = {mf.family_id: mf for mf in mfs}
-    npl._spec_lookup = {spec.spectrum_id: spec for spec in spectra}
-    # tmp path to store 'metcalf/metcalf_scores.pckl' file
-    # Must use `tmp_path_factory` (session scope) instead of `tmp_path` (function scope)
-    npl._loader._root = tmp_path_factory.mktemp('npl_test')
-    return npl
-
-
-@fixture(scope='module')
-def mc(npl) -> MetcalfScoring:
-    """MetcalfScoring object."""
-    mc = MetcalfScoring(npl)
-    mc.setup(npl)
-    return mc
 
 
 def test_get_links_gcf_standardised_false(npl, mc, gcfs, spectra, mfs,
